BAP-11039: Integrity constraint violation during IMAP synchronization

+++ src/Oro/Bundle/ImapBundle/Sync/ImapEmailSynchronizationProcessor.php
@@ -148,9 +148,15 @@ class ImapEmailSynchronizationProcessor extends AbstractEmailSynchronizationProc
                     /* @var $staleImapEmails ImapEmail[] */
                     foreach ($staleImapEmails as $imapEmail) {
                         $email = $imapEmail->getEmail();
-                        $email->getEmailUsers()->forAll(function ($key, EmailUser $emailUser) use ($folder) {
+                        $email->getEmailUsers()->forAll(function ($key, EmailUser $emailUser) use (
+                            $folder,
+                            $imapEmail
+                        ) {
+                            $existsEmails = $this->em->getRepository('OroImapBundle:ImapEmail')
+                                ->findBy(['email' => $imapEmail->getEmail()]);
+
                             $emailUser->removeFolder($folder);
-                            if (!$emailUser->getFolders()->count()) {
+                            if (count($existsEmails) <= 1 && !$emailUser->getFolders()->count()) {
                                 $this->em->remove($emailUser);
                             }
                         });
